ai outside query and print errors

+++ Curves/delete_duplicate_curves.py
@@ -19,7 +19,7 @@ def is_winter_hour_change(dt):
 def get_mongo_name_datetime_duplicateds(mongo_db, mongo_collection):
     data = dict()
     name_datetime_duplicateds_query = [
-        {'$group': {'_id': {'ai': '$ai', 'datetime': '$datetime', 'name': '$name'},
+        {'$group': {'_id': {'datetime': '$datetime', 'name': '$name'},
             'count': {'$sum': 1},
             'uniqueIds': {'$addToSet': '$_id'}}},
         {'$match': {'count': {'$gt': 1}}}
@@ -36,18 +36,26 @@ def treat_duplicateds(mongo_db, mongo_collection, doit=False):
     total_deleted = 0
     duplicateds = get_mongo_name_datetime_duplicateds(mongo_db, mongo_collection)
     for entry in tqdm(duplicateds):
-        if not 'name' in entry['_id'] or 'ai' not in entry['_id']:
-            continue
-        to_delete = not is_winter_hour_change(entry['_id']['datetime'])
-        if to_delete:
-            cr = list(mongo_db[mongo_collection].find({'_id':{'$in': entry['uniqueIds']}}))
-            informed_ai = [x['ai'] for x in cr]
-            if len(set(informed_ai)) == 1:
-                total_deleted + = len(cr)-1
-                if doit:
-                    del_res = mongo_db[mongo_collection].delete_many({'_id':{'$in': entry['uniqueIds'][1:]}})
-            else:
-                warn("Repetits diferents ", cr)
+        try:
+            if not 'name' in entry['_id'] or 'ai' not in entry['_id']:
+                continue
+            to_delete = not is_winter_hour_change(entry['_id']['datetime'])
+            if to_delete:
+                cr = list(mongo_db[mongo_collection].find({'_id':{'$in': entry['uniqueIds']}}))
+                informed_ai = [x['ai'] for x in cr if 'ai' in x]
+                if len(informed_ai) != len(cr):
+                    warn("Element/s sense ai", cr)
+                    continue
+                if len(set(informed_ai)) == 1:
+                    total_deleted + = len(cr)-1
+                    if doit:
+                        del_res = mongo_db[mongo_collection].delete_many({'_id':{'$in': entry['uniqueIds'][1:]}})
+                else:
+                    warn("Repetits diferents ", cr)
+        except KeyboardInterrupt as e:
+            break
+        except Exception as e:
+            error("Error: {}".format(e))
     success("Eliminats {} registres".format(total_deleted))
 
 def main():
